+++ doorstop/cli/main.py
@@ -199,7 +199,7 @@ def _run(args, cwd, err):  # pylint: disable=W0613
     """
     try:
         tree = build(cwd)
-        valid = tree.check()
+        valid = tree.valid()
     except DoorstopError as error:
         logging.error(error)
         return False