+++ doorstop/core/test/test_item.py
@@ -58,7 +58,7 @@ class TestItem(unittest.TestCase):  # pylint: disable=R0904
         self.assertEqual('', self.item._data['text'])
         self.assertEqual(set(), self.item._data['links'])
         self.assertEqual((1, 1, 1), self.item._data['level'])
-        self.assertTrue(self.item.check())
+        self.assertTrue(self.item.valid())
 
     def test_load_error(self):
         """Verify an exception is raised with invalid YAML."""
@@ -304,41 +304,42 @@ class TestItem(unittest.TestCase):  # pylint: disable=R0904
         self.assertRaises(DoorstopError,
                           Item.new, FILES, FILES, 'REQ', '', 2, 3, (1, 2, 3))
 
-    def test_check_invalid_ref(self):
-        """Verify an invalid reference fails check."""
+    def test_valid_invalid_ref(self):
+        """Verify an invalid reference fails valid."""
         with patch('doorstop.core.item.Item.find_ref',
                    Mock(side_effect=DoorstopError)):
-            self.assertFalse(self.item.check())
+            self.assertFalse(self.item.valid())
 
-    def test_check_inactive(self):
+    def test_valid_inactive(self):
         """Verify an inactive item is not checked."""
         self.item.active = False
-        self.item.ref = "invalid" + "reference"  # addition to avoid match
-        self.assertTrue(self.item.check())
+        with patch('doorstop.core.item.Item.find_ref',
+                   Mock(side_effect=DoorstopError)):
+            self.assertTrue(self.item.valid())
 
-    def test_check_link_to_inactive(self):
+    def test_valid_link_to_inactive(self):
         """Verify a link to an inactive item can be checked."""
         item = Mock()
         item.active = False
         tree = MagicMock()
         tree.find_item = Mock(return_value=item)
         self.item.links = ['a']
-        self.assertTrue(self.item.check(tree=tree))
+        self.assertTrue(self.item.valid(tree=tree))
 
-    def test_check_document(self):
+    def test_valid_document(self):
         """Verify an item can be checked against a document."""
         document = Mock()
         document.parent = 'fake'
-        self.assertTrue(self.item.check(document=document))
+        self.assertTrue(self.item.valid(document=document))
 
-    def test_check_document_with_links(self):
+    def test_valid_document_with_links(self):
         """Verify an item can be checked against a document with links."""
         self.item.add_link('unknown1')
         document = Mock()
         document.parent = 'fake'
-        self.assertTrue(self.item.check(document=document))
+        self.assertTrue(self.item.valid(document=document))
 
-    def test_check_tree(self):
+    def test_valid_tree(self):
         """Verify an item can be checked against a tree."""
 
         def mock_iter(self):  # pylint: disable=W0613
@@ -360,9 +361,9 @@ class TestItem(unittest.TestCase):  # pylint: disable=R0904
         tree = Mock()
         tree.__iter__ = mock_iter
         tree.find_item = lambda identifier: Mock(id='fake1')
-        self.assertTrue(self.item.check(tree=tree))
+        self.assertTrue(self.item.valid(tree=tree))
 
-    def test_check_tree_non_normative(self):
+    def test_valid_tree_non_normative(self):
         """Verify a non-normative item can be checked against a tree."""
 
         def mock_iter(self):  # pylint: disable=W0613
@@ -385,9 +386,9 @@ class TestItem(unittest.TestCase):  # pylint: disable=R0904
         tree = Mock()
         tree.__iter__ = mock_iter
         tree.find_item = lambda identifier: Mock(id='fake1')
-        self.item.check(tree=tree)
+        self.item.valid(tree=tree)
 
-    def test_check_tree_no_down_links(self):
+    def test_valid_tree_no_down_links(self):
         """Verify an item can be checked against a tree without down links."""
 
         def mock_iter(self):  # pylint: disable=W0613
@@ -409,14 +410,14 @@ class TestItem(unittest.TestCase):  # pylint: disable=R0904
         tree = Mock()
         tree.__iter__ = mock_iter
         tree.find_item = lambda identifier: Mock(id='fake1')
-        self.assertTrue(self.item.check(tree=tree))
+        self.assertTrue(self.item.valid(tree=tree))
 
-    def test_check_tree_error(self):
+    def test_valid_tree_error(self):
         """Verify an item can be checked against a tree with errors."""
         self.item.add_link('fake1')
         tree = MagicMock()
         tree.find_item = Mock(side_effect=DoorstopError)
-        self.assertFalse(self.item.check(tree=tree))
+        self.assertFalse(self.item.valid(tree=tree))
 
     @patch('os.remove')
     def test_delete(self, mock_remove):